Convert hard coded html to JavaScript coded
Now that the webpage was designed and hardcoded in the index.html file,
I removed all of that hard code to be converted into JavaScript in the
src/index.js file. The Styling however was made into a separate css file
which was added to the src file. The src/index.js file makes use of
the import keywords used with all necessary assets (images and the
style.css file), and it makes use of webpack to convert the src/index.js
file, with all of its assets, into a dist/main.js file. The
webpack.config.js file was modified, in tandem, to look for style files,
images, and fonts. Finally, the package.json file was modified so that
webpack runs automatically when running npm.