Create menu and contact tabs and tab logic
I created menu and contact tabs, which are separate JavaScript files.
Each script file contains a function module that is exported to the
index.js file, which contains the logic for accessing the different
tabs. I also organized my scripts into their own folder, and edited the
webpack config file and other files to take that into consideration.
Finally, I had originally created a css file for each tab, however I
realized that this is difficult to implement because when you import a
style sheet the the index.js file, it is automatically implemented. I
therefor reduced the amount of style sheets to one.